Update for Lib. v2022.1 - All readmes updated, no. of IRQs shown fixed
The readme file:./Libero_Projects/README.md was showing the wrong
number of 'Internal IRQs' for the 3 configurations available for
MIV_RV32. The number was changed from 6 to 1. The designs used
to support 6 'Internal IRQs' up until v12.6 release of the scripts,
when that number was reduced to 1 as, no more interrupts were needed.

The readmes have been updated with information on the new 3rd
argument.It was implemented to allow design building targeted for
an 'ES' die type. Other minor fixes were made to the readme files.

All readmes have been updated to now show that the latest Libero SoC
v2022.1 is supported.

Author: Sebastian Slowikowski <seb.slowikowski@microchip.com>
Signed-off: Sebastian Slowikowski <seb.slowikowski@microchip.com>